requirements.txt
anyio==4.6.2.post1
click==8.1.7
colorama==0.4.6
fastapi==0.95.0
h11==0.14.0
idna==3.10
imageio==2.36.0
lazy_loader==0.4
networkx==3.4.2
numpy==1.23.5
opencv-python-headless==4.5.5.64
packaging==24.1
Pillow==9.4.0
pydantic==1.10.18
python-multipart==0.0.17
scikit-image==0.24.0
scipy==1.14.1
sniffio==1.3.1
starlette==0.26.1
tifffile==2024.9.20
typing_extensions==4.12.2
uvicorn==0.21.1
